feat(TDP-5575) : Make the multitenant users for IT configurable * First functional code.
