+++ dataprep-test-api/src/test/java/org/talend/dataprep/qa/config/BackendProperties.java
@@ -1,16 +1,16 @@
 /*
- *  ============================================================================
+ * ============================================================================
  *
- *  Copyright (C) 2006-2018 Talend Inc. - www.talend.com
+ * Copyright (C) 2006-2018 Talend Inc. - www.talend.com
  *
- *  This source code is available under agreement available at
- *  https://github.com/Talend/data-prep/blob/master/LICENSE
+ * This source code is available under agreement available at
+ * https://github.com/Talend/data-prep/blob/master/LICENSE
  *
- *  You should have received a copy of the agreement
- *  along with this program; if not, write to Talend SA
- *  9 rue Pages 92150 Suresnes, France
+ * You should have received a copy of the agreement
+ * along with this program; if not, write to Talend SA
+ * 9 rue Pages 92150 Suresnes, France
  *
- *  ============================================================================
+ * ============================================================================
  */
 
 package org.talend.dataprep.qa.config;
@@ -24,30 +24,16 @@ public class BackendProperties {
 
     private API api = new API();
 
-    public static class User {
+    public User getUser() {
+        return user;
+    }
 
-        private String login;
-
-        private String password;
-
-        public String getLogin() {
-            return login;
-        }
-
-        public void setLogin(String login) {
-            this.login = login;
-        }
-
-        public String getPassword() {
-            return password;
-        }
-
-        public void setPassword(String password) {
-            this.password = password;
-        }
+    public API getApi() {
+        return api;
     }
 
     public static class API {
+
         private String url;
 
         public String getUrl() {
@@ -57,13 +43,5 @@ public class BackendProperties {
         public void setUrl(String url) {
             this.url = url;
         }
-    }
-
-    public User getUser() {
-        return user;
-    }
-
-    public API getApi() {
-        return api;
     }
 }
